refactor: decorator: Unify logic between @OneToOne/@OneToMany/@ManyToOne/@ManyToMany
