+++ src/decorator/relations/ManyToMany.ts
@@ -1,12 +1,12 @@
-import {getMetadataArgsStorage, ObjectType, RelationOptions} from "../../";
-import {RelationMetadataArgs} from "../../metadata-args/RelationMetadataArgs";
+import {ObjectType, RelationOptions} from "../../";
+import {Relation} from "./Relation";
 
 /**
  * Many-to-many is a type of relationship when Entity1 can have multiple instances of Entity2, and Entity2 can have
  * multiple instances of Entity1. To achieve it, this type of relation creates a junction table, where it storage
  * entity1 and entity2 ids. This is owner side of the relationship.
  */
-export function ManyToMany<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>), 
+export function ManyToMany<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>),
                               options?: RelationOptions): PropertyDecorator;
 
 /**
@@ -26,35 +26,5 @@ export function ManyToMany<T>(typeFunctionOrTarget: string|((type?: any) => Obje
 export function ManyToMany<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>),
                               inverseSideOrOptions?: string|((object: T) => any)|RelationOptions,
                               options?: RelationOptions): PropertyDecorator {
-
-    // normalize parameters
-    let inverseSideProperty: string|((object: T) => any);
-    if (typeof inverseSideOrOptions === "object") {
-        options = <RelationOptions> inverseSideOrOptions;
-    } else {
-        inverseSideProperty = <string|((object: T) => any)> inverseSideOrOptions;
-    }
-
-    return function (object: Object, propertyName: string) {
-        if (!options) options = {} as RelationOptions;
-
-        // now try to determine it its lazy relation
-        let isLazy = options.lazy === true;
-        if (!isLazy && Reflect && (Reflect as any).getMetadata) { // automatic determination
-            const reflectedType = (Reflect as any).getMetadata("design:type", object, propertyName);
-            if (reflectedType && typeof reflectedType.name === "string" && reflectedType.name.toLowerCase() === "promise")
-                isLazy = true;
-        }
-
-        getMetadataArgsStorage().relations.push({
-            target: object.constructor,
-            propertyName: propertyName,
-            // propertyType: reflectedType,
-            relationType: "many-to-many",
-            isLazy: isLazy,
-            type: typeFunctionOrTarget,
-            inverseSideProperty: inverseSideProperty,
-            options: options
-        } as RelationMetadataArgs);
-    };
+    return Relation<T>("many-to-many", typeFunctionOrTarget, inverseSideOrOptions, options);
 }
