+++ src/decorator/relations/ManyToOne.ts
@@ -1,5 +1,5 @@
-import {getMetadataArgsStorage, ObjectType, RelationOptions} from "../../";
-import {RelationMetadataArgs} from "../../metadata-args/RelationMetadataArgs";
+import {ObjectType, RelationOptions} from "../../";
+import {Relation} from "./Relation";
 
 /**
  * A many-to-one relation allows creating the type of relation where Entity1 can have a single instance of Entity2, but
@@ -26,34 +26,5 @@ export function ManyToOne<T>(typeFunctionOrTarget: string|((type?: any) => Objec
                              inverseSideOrOptions?: string|((object: T) => any)|RelationOptions,
                              options?: RelationOptions): PropertyDecorator {
 
-    // Normalize parameters.
-    let inverseSideProperty: string|((object: T) => any);
-    if (typeof inverseSideOrOptions === "object") {
-        options = <RelationOptions> inverseSideOrOptions;
-    } else {
-        inverseSideProperty = <string|((object: T) => any)> inverseSideOrOptions;
-    }
-
-    return function (object: Object, propertyName: string) {
-        if (!options) options = {} as RelationOptions;
-
-        // Now try to determine if it is a lazy relation.
-        let isLazy = options && options.lazy === true ? true : false;
-        if (!isLazy && Reflect && (Reflect as any).getMetadata) { // automatic determination
-            const reflectedType = (Reflect as any).getMetadata("design:type", object, propertyName);
-            if (reflectedType && typeof reflectedType.name === "string" && reflectedType.name.toLowerCase() === "promise")
-                isLazy = true;
-        }
-
-        getMetadataArgsStorage().relations.push({
-            target: object.constructor,
-            propertyName: propertyName,
-            // propertyType: reflectedType,
-            relationType: "many-to-one",
-            isLazy: isLazy,
-            type: typeFunctionOrTarget,
-            inverseSideProperty: inverseSideProperty,
-            options: options
-        } as RelationMetadataArgs);
-    };
+    return Relation<T>("many-to-one", typeFunctionOrTarget, inverseSideOrOptions, options);
 }
