+++ src/decorator/relations/OneToOne.ts
@@ -1,11 +1,11 @@
-import {getMetadataArgsStorage, ObjectType, RelationOptions} from "../../";
-import {RelationMetadataArgs} from "../../metadata-args/RelationMetadataArgs";
+import {ObjectType, RelationOptions} from "../../";
+import {Relation} from "./Relation";
 
 /**
  * One-to-one relation allows to create direct relation between two entities. Entity1 have only one Entity2.
  * Entity1 is an owner of the relationship, and storages Entity1 id on its own side.
  */
-export function OneToOne<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>), 
+export function OneToOne<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>),
                             options?: RelationOptions): PropertyDecorator;
 
 /**
@@ -23,35 +23,5 @@ export function OneToOne<T>(typeFunctionOrTarget: string|((type?: any) => Object
 export function OneToOne<T>(typeFunctionOrTarget: string|((type?: any) => ObjectType<T>),
                             inverseSideOrOptions?: string|((object: T) => any)|RelationOptions,
                             options?: RelationOptions): PropertyDecorator {
-
-    // normalize parameters
-    let inverseSideProperty: string|((object: T) => any);
-    if (typeof inverseSideOrOptions === "object") {
-        options = <RelationOptions> inverseSideOrOptions;
-    } else {
-        inverseSideProperty = <string|((object: T) => any)> inverseSideOrOptions;
-    }
-
-    return function (object: Object, propertyName: string) {
-        if (!options) options = {} as RelationOptions;
-
-        // now try to determine it its lazy relation
-        let isLazy = options && options.lazy === true ? true : false;
-        if (!isLazy && Reflect && (Reflect as any).getMetadata) { // automatic determination
-            const reflectedType = (Reflect as any).getMetadata("design:type", object, propertyName);
-            if (reflectedType && typeof reflectedType.name === "string" && reflectedType.name.toLowerCase() === "promise")
-                isLazy = true;
-        }
-
-        getMetadataArgsStorage().relations.push({
-            target: object.constructor,
-            propertyName: propertyName,
-            // propertyType: reflectedType,
-            isLazy: isLazy,
-            relationType: "one-to-one",
-            type: typeFunctionOrTarget,
-            inverseSideProperty: inverseSideProperty,
-            options: options
-        } as RelationMetadataArgs);
-    };
+    return Relation<T>("one-to-one", typeFunctionOrTarget, inverseSideOrOptions, options);
 }
